ASSUME_ALIGNED that works with icc and gcc.

+++ Matriplex/MatriplexCommon.h
@@ -48,7 +48,7 @@ namespace Matriplex
 
    inline void align_check(const char* pref, void *adr)
    {
-      printf("%s 0x%llx  -  modulo 64 = %d\n", pref, adr, (long long)adr%64);
+      printf("%s 0x%llx  -  modulo 64 = %lld\n", pref, (long long unsigned)adr, (long long)adr%64);
    }
 }
